Clean-up Cache In-Memory (#24151)
