+++ aspnetcore/performance/caching/memory/samples/3.x/RPcache/Services/MyMemoryCache.cs
@@ -2,7 +2,7 @@
 
 namespace RPcache.Services
 {
-    #region snippet
+    // <snippet>
     // using Microsoft.Extensions.Caching.Memory;
     public class MyMemoryCache 
     {
@@ -15,5 +15,5 @@ namespace RPcache.Services
             });
         }
     }
-    #endregion
+    // </snippet>
 }
