--- aspnetcore/performance/caching/memory/samples/6.0/CachingMemorySample/Snippets/Program.cs
@@ -1,15 +1,0 @@
-﻿namespace CachingMemorySample.Snippets
-{
-    public static class Program
-    {
-        public static void AddSingletonMyMemoryCache(string[] args)
-        {
-            #region snippet_AddSingletonMyMemoryCache
-            var builder = WebApplication.CreateBuilder(args);
-
-            // Add services to the container.
-            builder.Services.AddSingleton<MyMemoryCache>();
-            #endregion
-        }
-    }
-}
