+++ aspnetcore/performance/caching/memory/samples/6.x/CachingMemorySample/CacheKeys.cs
@@ -1,0 +1,16 @@
+﻿namespace CachingMemorySample;
+    
+public static class CacheKeys
+{
+    public const string Entry = "_Entry";
+
+    public const string CallbackEntry = "_CallbackEntry";
+
+    public const string CallbackMessage = "_CallbackMessage";
+
+    public const string Parent = "_Parent";
+
+    public const string Child = "_Child";
+
+    public const string DependentCancellationTokenSource = "_DependentCancellationTokenSource";
+}
