+++ aspnetcore/performance/caching/memory/samples/6.x/CachingMemorySample/Pages/Shared/_Layout.cshtml
