+++ aspnetcore/performance/caching/memory/samples/6.x/CachingMemorySample/Pages/_ViewStart.cshtml
