+++ aspnetcore/performance/caching/memory/samples/6.x/CachingMemorySample/appsettings.Development.json
