+++ aspnetcore/performance/caching/memory/samples/6.x/CachingMemorySample/wwwroot/favicon.ico
